Feature optional update (#96)
* Update system.FileSystemError to Go 1.13 error semantics

* Move FileSystemError definition to own file

* Add variant of IsPermission() which is compatible with go 1.13 nested error semantics

* Add concept of optional/mandatory systembundle updates

* Move DownloadStatusPanel to own file

* Implement pausing logic.

* colored progress bars on Windows

* Retain window size after resuming pause on Windows

* Fix displayed progress being too low when some stages were skipped

* Refactor run.go

* Eliminate duplicate dependency injection

* Eliminate potential deadlock in case user manages to click Continue twice

* Reduce maxLineWidth to 110 characters to more closely align with initial window width

* Fix empty LauncherUpdate causing error despite documentation explicitly allowing it

* Set progress bar pause state during stages of active waiting

* Implement common Aero progress bar bug workaround

* Make sure inapplicable systembundle differences are logged

* fix false log warnings about diverged program name

* Add IsUpdateMandatory for Bundles. Not adding it for LauncherUpdate: weird middle-ground between system and normal mode aka feature creep

+++ pkg/launcher/bundle/update_bundle.go
@@ -20,16 +20,11 @@ import (
 	"github.com/setlog/trivrost/pkg/launcher/hashing"
 )
 
-func (u *Updater) DetermineBundleUpdateRequired(userBundlesFolderPath, systemBundlesFolderPath string) bool {
+func (u *Updater) DetermineBundleRequirements(userBundlesFolderPath, systemBundlesFolderPath string) {
 	u.userBundlesFolderPath, u.systemBundlesFolderPath = userBundlesFolderPath, systemBundlesFolderPath
 	u.determineLocalBundleVersions()
 	u.removeUnknownBundles()
 	u.determineBundleChanges()
-	isUpdatedRequired := u.isAtLeastOneChangeRequired()
-	if isUpdatedRequired {
-		u.assertUpdatePossible()
-	}
-	return isUpdatedRequired
 }
 
 func (u *Updater) InstallBundleUpdates() {
@@ -123,11 +118,18 @@ func (u *Updater) RetrieveBundleInfo(fromURL string, publicKeys []*rsa.PublicKey
 func (u *Updater) installBundleUpdates() {
 	u.announceStatus(DownloadBundleFiles, countUpdatesBytes(u.bundleUpdateInfos))
 	for _, bundleUpdateConfig := range u.bundleUpdateInfos {
-		log.Infof("Downloading %d files for bundle \"%s\".", bundleUpdateConfig.WantedState.UpdateFileCount(), bundleUpdateConfig.LocalDirectory)
-		bundleDirectory := filepath.Join(u.userBundlesFolderPath, bundleUpdateConfig.LocalDirectory)
-		deleteChangedFiles(bundleUpdateConfig.WantedState, bundleDirectory)
-		u.downloader.MustDownloadToDirectory(bundleUpdateConfig.BaseURL, bundleUpdateConfig.WantedState, bundleDirectory)
-		system.MustRecursivelyRemoveEmptyFolders(bundleDirectory)
+		if bundleUpdateConfig.IsSystemBundle {
+			if bundleUpdateConfig.WantedState.HasChanges() {
+				log.Warnf("Cannot update bundle \"%s\" because it is a system bundle. The following changes will not be applied:", bundleUpdateConfig.LocalDirectory)
+				bundleUpdateConfig.LogChanges()
+			}
+		} else {
+			log.Infof("Downloading %d files for bundle \"%s\".", bundleUpdateConfig.WantedState.UpdateFileCount(), bundleUpdateConfig.LocalDirectory)
+			bundleDirectory := filepath.Join(u.userBundlesFolderPath, bundleUpdateConfig.LocalDirectory)
+			deleteChangedFiles(bundleUpdateConfig.WantedState, bundleDirectory)
+			u.downloader.MustDownloadToDirectory(bundleUpdateConfig.BaseURL, bundleUpdateConfig.WantedState, bundleDirectory)
+			system.MustRecursivelyRemoveEmptyFolders(bundleDirectory)
+		}
 	}
 }
